fix: revert accidental swap
# Conflicts:
#	app/src/main/java/com/github/se/cyrcle/ui/profile/ViewProfileScreen.kt
